+++ include/proofseed/planting.h
@@ -27,21 +27,18 @@
 #ifndef PROOFSEED_PLANTING_H
 #define PROOFSEED_PLANTING_H
 
-#include "proofseed/future.h"
+#include "proofseed/asynqro_extra.h"
 #include "proofseed/proofalgorithms.h"
 #include "proofseed/tasks.h"
 
 namespace algorithms = Proof::algorithms;
 namespace tasks = Proof::tasks;
+namespace futures = Proof::futures;
 
 using Proof::CancelableFuture;
 using Proof::Failure;
 using Proof::Future;
-using Proof::FutureSP;
-using Proof::FutureWP;
 using Proof::Promise;
-using Proof::PromiseSP;
-using Proof::PromiseWP;
 using Proof::WithFailure;
 
 #endif // PROOFSEED_PLANTING_H