+++ include/proofseed/tasks.h
@@ -27,125 +27,35 @@
 #ifndef PROOF_TASKS_H
 #define PROOF_TASKS_H
 
-#include "proofseed/future.h"
-#include "proofseed/proofalgorithms.h"
 #include "proofseed/proofseed_global.h"
 
-#include "3rdparty/optional/optional.hpp"
-
 #include <QEventLoop>
-#include <QVector>
+#include <QSharedPointer>
 
-#include <cmath>
-#include <type_traits>
+#include <functional>
 
 namespace Proof {
 namespace tasks {
 
-enum class RestrictionType
+class PROOF_SEED_EXPORT TasksExtra
 {
-    Custom,
-    Intensive,
-    ThreadBound
-};
-
-class TasksDispatcherPrivate;
-class Worker;
-class PROOF_SEED_EXPORT TasksDispatcher
-{
-    Q_DECLARE_PRIVATE(TasksDispatcher)
 public:
-    static TasksDispatcher *instance();
-    TasksDispatcher(const TasksDispatcher &) = delete;
-    TasksDispatcher(TasksDispatcher &&) = delete;
-    TasksDispatcher &operator=(const TasksDispatcher &) = delete;
-    TasksDispatcher &operator=(TasksDispatcher &&) = delete;
-
-    qint32 capacity() const;
-    qint32 restrictorCapacity(RestrictionType restrictionType, const QString &restrictor = QString()) const;
-    void setCapacity(qint32 capacity);
-    void addCustomRestrictor(const QString &restrictor, qint32 capacity);
-
-    template <typename Task, typename Result = typename std::result_of_t<Task()>,
-              typename = typename std::enable_if_t<!std::is_same<Result, void>::value
-                                                   && !IsSpecialization<Result, QSharedPointer, Future>::value>>
-    CancelableFuture<Result> run(Task &&task, RestrictionType restrictionType, const QString &restrictor) noexcept
-    {
-        PromiseSP<Result> promise = PromiseSP<Result>::create();
-        std::function<void()> f = [promise, task = std::forward<Task>(task)]() noexcept
-        {
-            if (promise->filled())
-                return;
-            Proof::futures::detail::resetLastFailure();
-            try {
-                promise->success(task());
-            } catch (const std::exception &e) {
-                promise->failure(Failure::fromException(e));
-            } catch (...) {
-                promise->failure(Failure::fromException());
-            }
-        };
-        insertTaskInfo(std::move(f), restrictionType, restrictor);
-        return CancelableFuture<Result>(promise);
-    }
-
-    template <typename Task, typename Result = typename std::result_of_t<Task()>,
-              typename InnerResult = typename Result::Type::Value,
-              typename = typename std::enable_if_t<IsSpecialization<Result, QSharedPointer, Future>::value>>
-    CancelableFuture<InnerResult> run(Task &&task, RestrictionType restrictionType, const QString &restrictor) noexcept
-    {
-        PromiseSP<InnerResult> promise = PromiseSP<InnerResult>::create();
-        std::function<void()> f = [promise, task = std::forward<Task>(task)]() noexcept
-        {
-            if (promise->filled())
-                return;
-            Proof::futures::detail::resetLastFailure();
-            try {
-                task()
-                    ->onSuccess([promise](const InnerResult &result) noexcept { promise->success(result); })
-                    ->onFailure([promise](const Failure &failure) noexcept { promise->failure(failure); });
-            } catch (const std::exception &e) {
-                promise->failure(Failure::fromException(e));
-            } catch (...) {
-                promise->failure(Failure::fromException());
-            }
-        };
-        insertTaskInfo(std::move(f), restrictionType, restrictor);
-        return CancelableFuture<InnerResult>(promise);
-    }
-
-    template <typename Task, typename Result = typename std::result_of_t<Task()>,
-              typename = typename std::enable_if_t<std::is_same<Result, void>::value>>
-    CancelableFuture<bool> run(Task &&task, RestrictionType restrictionType, const QString &restrictor) noexcept
-    {
-        PromiseSP<bool> promise = PromiseSP<bool>::create();
-        std::function<void()> f = [promise, task = std::forward<Task>(task)]() noexcept
-        {
-            if (promise->filled())
-                return;
-            Proof::futures::detail::resetLastFailure();
-            try {
-                task();
-                promise->success(true);
-            } catch (const std::exception &e) {
-                promise->failure(Failure::fromException(e));
-            } catch (...) {
-                promise->failure(Failure::fromException());
-            }
-        };
-        insertTaskInfo(std::move(f), restrictionType, restrictor);
-        return CancelableFuture<bool>(promise);
-    }
+    TasksExtra() = delete;
+    TasksExtra(const TasksExtra &) = delete;
+    TasksExtra(TasksExtra &&) = delete;
+    TasksExtra &operator=(const TasksExtra &) = delete;
+    TasksExtra &operator=(TasksExtra &&) = delete;
+    ~TasksExtra() = delete;
 
     template <class SignalSender, class SignalType, class... Args>
-    void addSignalWaiter(SignalSender *sender, SignalType signal, std::function<bool(Args...)> callback) noexcept
+    static void addSignalWaiter(SignalSender *sender, SignalType signal, std::function<bool(Args...)> callback) noexcept
     {
-        std::function<void(const QSharedPointer<QEventLoop> &)> connector = [sender, signal, callback, this](
+        std::function<void(const QSharedPointer<QEventLoop> &)> connector = [sender, signal, callback](
             const QSharedPointer<QEventLoop> &eventLoop) noexcept
         {
             auto connection = QSharedPointer<QMetaObject::Connection>::create();
             auto eventLoopWeak = eventLoop.toWeakRef();
-            std::function<void(Args...)> slot = [callback, eventLoopWeak, connection, this](Args... args) {
+            std::function<void(Args...)> slot = [callback, eventLoopWeak, connection](Args... args) {
                 QSharedPointer<QEventLoop> eventLoop = eventLoopWeak.toStrongRef();
                 if (!eventLoop)
                     return;
@@ -162,189 +72,23 @@ public:
         addSignalWaiterPrivate(std::move(connector));
     }
 
-    void fireSignalWaiters() noexcept;
+    static void fireSignalWaiters() noexcept;
 
 private:
-    friend class Worker;
-    TasksDispatcher();
-    ~TasksDispatcher();
-    void insertTaskInfo(std::function<void()> &&wrappedTask, RestrictionType restrictionType,
-                        const QString &restrictor) noexcept;
-    void addSignalWaiterPrivate(std::function<void(const QSharedPointer<QEventLoop> &)> &&connector) noexcept;
-    bool eventLoopStarted() const noexcept;
-    void clearEventLoop() noexcept;
-
-    QScopedPointer<TasksDispatcherPrivate> d_ptr;
+    static void addSignalWaiterPrivate(std::function<void(const QSharedPointer<QEventLoop> &)> &&connector) noexcept;
+    static bool eventLoopStarted() noexcept;
+    static void clearEventLoop() noexcept;
 };
-
-template <typename Task>
-auto run(Task &&task, RestrictionType restrictionType = RestrictionType::Custom,
-         const QString &restrictor = QString()) noexcept
-    -> decltype(TasksDispatcher::instance()->run(task, restrictionType, restrictor))
-{
-    return TasksDispatcher::instance()->run(std::forward<Task>(task), restrictionType, restrictor);
-}
-
-template <typename Task>
-auto run(RestrictionType restrictionType, const QString &restrictor, Task &&task) noexcept
-    -> decltype(TasksDispatcher::instance()->run(task, restrictionType, restrictor))
-{
-    return TasksDispatcher::instance()->run(std::forward<Task>(task), restrictionType, restrictor);
-}
-
-template <template <typename...> class Container, typename Input, typename Task,
-          typename Output = typename std::result_of_t<Task(Input)>,
-          typename = typename std::enable_if_t<!std::is_same<Output, void>::value
-                                               && !IsSpecialization<Output, QSharedPointer, Future>::value>>
-auto run(const Container<Input> &data, Task &&task, RestrictionType restrictionType = RestrictionType::Custom,
-         const QString &restrictor = QString()) noexcept
-    -> decltype(task(*(data.cbegin())), FutureSP<Container<Output>>())
-{
-    if (!data.size())
-        return Future<Container<Output>>::successful();
-    auto seq = algorithms::map(data,
-                               [task = std::forward<Task>(task), restrictionType,
-                                restrictor](const Input &x) -> FutureSP<Output> {
-                                   return run([x, task]() -> Output { return task(x); }, restrictionType, restrictor);
-                               });
-    return Future<>::sequence(seq);
-}
-
-template <template <typename...> class Container, typename Input, typename Task,
-          typename Output = typename std::result_of_t<Task(Input)>,
-          typename = typename std::enable_if_t<IsSpecialization<Output, QSharedPointer, Future>::value>,
-          typename OutputValue = typename Output::Type::Value>
-auto run(const Container<Input> &data, Task &&task, RestrictionType restrictionType = RestrictionType::Custom,
-         const QString &restrictor = QString()) noexcept
-    -> decltype(task(*(data.cbegin())), FutureSP<Container<OutputValue>>())
-{
-    if (!data.size())
-        return Future<Container<OutputValue>>::successful();
-    auto seq = algorithms::map(data,
-                               [task = std::forward<Task>(task), restrictionType, restrictor](const Input &x) -> Output {
-                                   return run([x, task]() -> Output { return task(x); }, restrictionType, restrictor);
-                               });
-    return Future<>::sequence(seq);
-}
-
-template <template <typename...> class Container, typename Input, typename Task,
-          typename Output = typename std::result_of_t<Task(Input)>,
-          typename = typename std::enable_if_t<std::is_same<Output, void>::value>>
-auto run(const Container<Input> &data, Task &&task, RestrictionType restrictionType = RestrictionType::Custom,
-         const QString &restrictor = QString()) -> decltype(task(*(data.cbegin())), FutureSP<bool>())
-{
-    if (!data.size())
-        return Future<>::successful(true);
-    auto seq = algorithms::map(data,
-                               [task = std::forward<Task>(task), restrictionType,
-                                restrictor](const Input &x) -> FutureSP<bool> {
-                                   return run([x, task]() { task(x); }, restrictionType, restrictor);
-                               });
-    return Future<>::sequence(seq)->andThenValue(true);
-}
-
-template <template <typename...> class Container, typename Input, typename Task,
-          typename Output = typename std::result_of_t<Task(long long, Input)>,
-          typename = typename std::enable_if_t<!std::is_same<Output, void>::value
-                                               && !IsSpecialization<Output, QSharedPointer, Future>::value>>
-auto run(const Container<Input> &data, Task &&task, RestrictionType restrictionType = RestrictionType::Custom,
-         const QString &restrictor = QString()) noexcept
-    -> decltype(task(0ll, *(data.cbegin())), FutureSP<Container<Output>>())
-{
-    if (!data.size())
-        return Future<Container<Output>>::successful();
-    auto seq = algorithms::map(data,
-                               [task = std::forward<Task>(task), restrictionType,
-                                restrictor](long long index, const Input &x) -> FutureSP<Output> {
-                                   return run([index, x, task]() -> Output { return task(index, x); }, restrictionType,
-                                              restrictor);
-                               });
-    return Future<>::sequence(seq);
-}
-
-template <template <typename...> class Container, typename Input, typename Task,
-          typename Output = typename std::result_of_t<Task(long long, Input)>,
-          typename = typename std::enable_if_t<IsSpecialization<Output, QSharedPointer, Future>::value>,
-          typename OutputValue = typename Output::Type::Value>
-auto run(const Container<Input> &data, Task &&task, RestrictionType restrictionType = RestrictionType::Custom,
-         const QString &restrictor = QString())
-    -> decltype(task(0ll, *(data.cbegin())), FutureSP<Container<OutputValue>>())
-{
-    if (!data.size())
-        return Future<Container<OutputValue>>::successful();
-    auto seq = algorithms::map(data,
-                               [task = std::forward<Task>(task), restrictionType,
-                                restrictor](long long index, const Input &x) -> Output {
-                                   return run([index, x, task]() -> Output { return task(index, x); }, restrictionType,
-                                              restrictor);
-                               });
-    return Future<>::sequence(seq);
-}
-
-template <template <typename...> class Container, typename Input, typename Task,
-          typename Output = typename std::result_of_t<Task(long long, Input)>,
-          typename = typename std::enable_if_t<std::is_same<Output, void>::value>>
-auto run(const Container<Input> &data, Task &&task, RestrictionType restrictionType = RestrictionType::Custom,
-         const QString &restrictor = QString()) noexcept -> decltype(task(0ll, *(data.cbegin())), FutureSP<bool>())
-{
-    if (!data.size())
-        return Future<>::successful(true);
-    auto seq = algorithms::map(data,
-                               [task = std::forward<Task>(task), restrictionType,
-                                restrictor](long long index, const Input &x) -> FutureSP<bool> {
-                                   return run([index, x, task]() { task(index, x); }, restrictionType, restrictor);
-                               });
-    return Future<>::sequence(seq)->andThenValue(true);
-}
-
-template <template <typename...> class Container, typename Input, typename Task,
-          typename Output = typename std::result_of_t<Task(Input)>>
-auto clusteredRun(const Container<Input> &data, Task &&task, qint64 minClusterSize = 1,
-                  RestrictionType restrictionType = RestrictionType::Custom, const QString &restrictor = QString()) noexcept
-    -> decltype((Container<Output>()).resize(data.count()), FutureSP<Container<Output>>())
-{
-    if (!data.size())
-        return Future<Container<Output>>::successful();
-    if (minClusterSize <= 0)
-        minClusterSize = 1;
-
-    return run([data, task = std::forward<Task>(task), minClusterSize, restrictionType, restrictor]() -> Container<Output> {
-        qint64 amount = data.count();
-        qint32 capacity = TasksDispatcher::instance()->restrictorCapacity(restrictionType, restrictor);
-        if (capacity == TasksDispatcher::instance()->capacity())
-            capacity = qMax(capacity / 2, TasksDispatcher::instance()->restrictorCapacity(RestrictionType::Intensive));
-        capacity = qMin(static_cast<qint32>(ceil(amount / static_cast<double>(minClusterSize))), capacity);
-        qint32 clusterSize = static_cast<qint32>(amount / capacity);
-        Container<Output> result;
-        result.resize(amount);
-        QVector<FutureSP<bool>> futures;
-        futures.reserve(capacity);
-        for (qint32 job = 0; job < capacity; ++job) {
-            qint32 right = job == (capacity - 1) ? amount : ((job + 1) * clusterSize + 1);
-            futures << run([&data, &task, &result, left = job * clusterSize, right]() {
-                for (qint32 i = left; i < right; ++i)
-                    result[i] = task(data[i]);
-            });
-        }
-        for (const auto &future : qAsConst(futures))
-            future->wait();
-        for (const auto &future : qAsConst(futures)) {
-            if (future->failed())
-                return WithFailure(future->failureReason());
-        }
-        return result;
-    });
-}
 
 template <typename SignalSender, typename SignalType, typename... Args>
 void addSignalWaiter(SignalSender *sender, SignalType signal, std::function<bool(Args...)> callback) noexcept
 {
-    TasksDispatcher::instance()->addSignalWaiter(sender, signal, callback);
+    TasksExtra::addSignalWaiter(sender, signal, callback);
 }
 
 inline void fireSignalWaiters() noexcept
 {
-    TasksDispatcher::instance()->fireSignalWaiters();
+    TasksExtra::fireSignalWaiters();
 }
 } // namespace tasks
 } // namespace Proof